allow for floating point parameters in the web interface and json

+++ esp32_config_webinterface_v7/webinterface.h
@@ -18,7 +18,7 @@
 /* these are for numbers */
 #define N_JSON_TO_CONFIG(x, y)   { if (root.containsKey(y)) { config.x = root[y]; } }
 #define N_CONFIG_TO_JSON(x, y)   { root[y] = config.x; }
-#define N_KEYVAL_TO_CONFIG(x, y) { if (server.hasArg(y))    { String str = server.arg(y); config.x = str.toInt(); } }
+#define N_KEYVAL_TO_CONFIG(x, y) { if (server.hasArg(y))    { String str = server.arg(y); config.x = str.toFloat(); } }
 
 /* these are for strings */
 #define S_JSON_TO_CONFIG(x, y)   { if (root.containsKey(y)) { strcpy(config.x, root[y]); } }
